% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vm_processing.R
\name{generate_sample_id_with_components}
\alias{generate_sample_id_with_components}
\title{Generate sample ID with components}
\usage{
generate_sample_id_with_components(
  site_code,
  parameter_name,
  environ_compartment,
  environ_compartment_sub,
  date,
  subsample = 1
)
}
\arguments{
\item{site_code}{Site code (vectorised)}

\item{parameter_name}{Parameter name (vectorised)}

\item{environ_compartment}{Environmental compartment (vectorised)}

\item{environ_compartment_sub}{Environmental sub-compartment (vectorised)}

\item{date}{Sampling date (vectorised)}

\item{subsample}{Subsample identifier (vectorised)}
}
\value{
Character vector of sample IDs in format:
{site_code}-{param_abbrev}-{comp_abbrev}-{date}-R-{subsample}
}
\description{
Creates unique sample identifiers by concatenating site code, parameter,
compartment, date, and subsample information. This is a helper function
copied from STOPeData::mod_samples_fct.R.
}
\details{
\itemize{
\item Parameter names are abbreviated to 8 characters (alphanumeric only)
\item Compartments are abbreviated to 12 characters (alphanumeric only)
\item Subsample values are truncated to 20 characters
}
}
\keyword{internal}
