% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vm_eData.R
\name{vm_create_edata_samples_table}
\alias{vm_create_edata_samples_table}
\title{Create eData samples table from intermediate samples-biota table}
\usage{
vm_create_edata_samples_table(vm_intermediate)
}
\arguments{
\item{vm_intermediate}{Intermediate samples-biota table from
vm_create_intermediate_samples_biota_table()}
}
\value{
A tibble conforming to eData samples schema
}
\description{
Extracts and validates the samples portion of the intermediate table,
conforming to the eData samples schema.
}
