% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vm_summary.R
\name{vm_matrix_class}
\alias{vm_matrix_class}
\title{Classify an eData Row's Matrix as Water / Sediment / Biota}
\usage{
vm_matrix_class(compartment, subcompartment)
}
\arguments{
\item{compartment}{\code{ENVIRON_COMPARTMENT}.}

\item{subcompartment}{\code{ENVIRON_COMPARTMENT_SUB}.}
}
\value{
Character vector, each element one of \code{"Water"}, \code{"Sediment"},
\code{"Biota"}.
}
\description{
Drives the \code{matrix} column and row ordering in \code{\link[=summarise_vm_dataset]{summarise_vm_dataset()}}.
Sediment is split out of the Aquatic compartment by sub-compartment;
anything neither biota nor sediment is Water.
}
